Capture trace when token is provided by browser (via beacon).
Rather than a boolean value in a cookie used to signal capturing a trace - a token identifier is passed into the agent from the beacon via the browser cookie. The agent captures a trace and echos the token identifier back to the RUM footer (and beacon). This allows the beacon to know which requests return.