Use Rubinius::GC numbers when available
These are way faster than using the agent. As anecdotal performance
metric, on an app I tested with with Newrelic the request time was
around 650ms, without it was 250ms.

Most of the time was spent in the Rubinius agent which Newrelic was
calling very often. This changes makes it perform on par with disabling
Newrelic in the test app.

I made the ordering prefer the Rubinius::GC here over the 1.9
statistics, which Rubinius also provides in 1.9 mode. These have to
implement some idiosynchrasies though to follow the MRI api so using
these directly is the faster approach.