Put in work-around for JRuby Dir.glob
There's a known issue with JRuby's Dir.glob() implementation:
https://github.com/jruby/jruby/issues/358 - where it isn't working properly
when the code is executed from within a jar. This is special because it doesn't
blow up - Dir.glob() simply returns the empty array: []. The fix is to remove
the 'jar:' from the front of the path name - if it exists. Simple, but
effective.