RUBY-961 Fix incorrect application of conditions in sinatra instrumentation
The agent's sinatra instrumentation was causing sinatra's condidtions to
be incorrectly applied in some obscure cases.  The bug was triggered
when a condition was present on a lower priority route that would match
the current request, except for the precense of a higher priority route.

For example /user/login would have the conditions for a /user/:id route
incorrectly applied.