RUBY-981 expose Marshaller#prepare, make double-encoding subtrees optional
For the audit log, we need to be able to generate a human-readable
representation of the object graph we are sending up to the server.
The JsonMarshaller interferes by double-encoding some parts of its
object graph as base64'd, gzipped JSON dumps.

This change makes that double-encoding of subtrees optional (by
passing a different encoder into Marshaller#prepare.