RUBY-984 audit log request bodies in JSON if marshaller is JSON
The audit log request bodies will now be in JSON if the JSON marshaller
is in use. Request bodies in the log will still be produced by inspect
if the Pruby marshaller is in use.

Final compression of the request body has been moved out of the marshaller
into the NewRelicService class itself to make this easier. I've justified
this to myself because:

1. compression of the final request body alters the content-encoding header,
   and thus feels like a transport-level concern. Marshallers should not
   know or care about the transport layer they are producing output for.
2. decompression of responses is handled in NewRelicService, not the marshallers,
   so the new approach is more symmetric in this regard.