RUBY-992 Don't report on pipe unless parent is already connected
In forked child processes that are setup to report to their parent process
via a PipeService, do not actually report anything and immediately send an
EOF on the pipe unless the parent process has already connected to the server
and received its config.

The purpose of doing this is to avoid accidentally collecting data that the
server config would have instructed us not to, and also not block the startup
of workers on the successful connection to the server.

This means there will be a brief window during which some Resque jobs may run
and not get any information recorded, but this is thought preferable to running
with a possibly incorrect config (i.e. with local config only) and accidentally
capturing sensitive information that the server config would have barred (esp.
given the relative infrequency with which most users restart Resque workers).