Added search paths for jar locations
The new way NewRelic looks for the config files is to have a series of paths -
with the first in the list being the highest priority, and then continuing
down from there. They start with a lot of good ones, but NONE of them are
found by jruby when running from a jar - but that's old news, isn't it? So I
added in the code to put the jar locations at the top of the list - if we're
running from a jar, and using the default warbler jar name/path scheme. If we
get off this beaten path, we might have to revisit this, but it's easy
enough to have multiples in the path - the first hit is the one that's used.