RUBY-1019 - Fix Webrick misconfiguration that messes up keep-alive
Webrick thought we wanted it to use nil as its access log.  It would
succeed on the first request, then experience an error writing the
access_log that looked like this:

ERROR NoMethodError: undefined method `+' for nil:NilClass
  webrick/httpserver.rb:213:in `block in access_log'
  webrick/httpserver.rb:212:in `each'
  webrick/httpserver.rb:212:in `access_log'
  webrick/httpserver.rb:111:in `run'
  webrick/server.rb:191:in `block in start_thread'

This error would screw up the TCP connection and subsequent requests on
that TCP connection would fail.  This was never a problem until we
implemented keep-alive.