RUBY-775 Handle locking of the StatsEngine backing hash in StatsEngine itself
This replaces the SynchronizedHash class with a new StatsHash class that knows
nothing about locking, pushing the responsibility for synchronization up to the
StatsEngine itself. Also, initialize the lock protecting the StatsHash instance
upon StatsEngine initialization (rather than lazily), to ensure that it's not
initialized twice (since ||= is not atomic).

Recording data to the StatsEngine is still not thread-safe, but this is a stepping
stone on the way there.