RUBY-775 Adjust record_metric API, and add increment_metric API
The public-facing record_metric API will not accept an options hash or a block.
If called from within the context of a traced transaction, it record both a
scoped (with the scope being the current transaction name) and unscoped metric.

To accomodate the relatively common increment-only use case, a new public-facing
increment_metric method has been added.