RUBY-775 Public custom metric recording APIs should not create scoped metrics
Recording to scoped metrics with the scope equal to the current transaction name
was considered, but decided against because of the fact that scoped metrics with
scope names equal to the transaction name would interfere with transaction
breakdown graphs. Specifically, custom metrics might not represent time-based
measurements, and even if they do, we have no current mechanism to cause their
timings to be subtracted from the exclusive time of the parent metrics.