RUBY-841 Automatically detect Resque as a dispatcher
With this change, it should no longer be necessary to explicitly set
NEWRELIC_DISPATCHER=resque in the environment when starting Resque workers
in order to get instrumentation.

The Resque check is more conservative than most in that it attempts to determine
not only that Resque's code is available, but also that we're actually
running in the context of a resque worker. Because of this, it must come
early in the list of candidate dispatchers (since the other dispatcher checks
are more liberal in their criteria and may trigger false positives).