ACDC-7 Test for checking license information
This commit adds two tests, and an update to the LICENSE file.

The first test will enforce that each ruby and javascript file includes
a short header describing its license.

The second will search through all files looking for references to
copyright, GPL, etc.  There is a list of known occurrences of these terms
which is honored.