Use info level for logging config file path instead of debug
This log message is most useful when debugging issues with the config file location.
In situations like this, it's difficult to bump up the log level because it's uncertain
whether the config file is being found / obeyed, so bumping this message to info
makes it possible to see the message at the default log level.