RUBY-1050 Avoid corrupting the scope stack when HTTP requests fail
When traced a trace HTTP request fails due to a network error, timeout,
or similar, we need to ensure that we still pop the scope stack so
that we can move on with tracing of the transaction. The failure may
be rescued by user code at a higher level, so the transaction might
continue without an error propagating.

Conflicts:
	lib/new_relic/agent/cross_app_tracing.rb