Tag seen exceptions with an ivar rather than checking object_ids
We tag exception objects to avoid duplicates, but there are a few
constraints:

1. We don't want to actually hold a reference to the exception objects
   to avoid unnecessarily retaining them.
2. We don't want the tagging to result in false positives (that is, we
   shouldn't incorrectly flag an exception as a duplicate if it isn't)
3. We want the tagging itself to be fast.

The previous approach satisfied constraints 1 and 3, but not 2, since
we were maintaining a list of seen object_ids (cleared periodically),
and object_ids can be reused in Ruby once the objects they're
associated with are GC'd. The ivar approach satisfies all three
constraints.