RUBY-977 Do not maintain long-lived DB connections for EXPLAIN queries
The original design of the explain plan feature involved us opening a
connection at harvest time, issuing EXPLAIN queries, then closing the
connection. This was changed in dd558a4 to avoid closing the DB
connection in forked web servers as a speculative fix for an issue
that turned out to be unrelated.

This change makes it so that we no longer predicate the closing of the
connection we manage for the purposes of EXPLAIN quereis on being in a
non-forked process.