RUBY-1062 Don't require a "detected dispatcher" to start
The agent has always attempted to detect whether it's running in a web
server or a background job (e.g. resque) before starting up.  This
requires custom detection code for each web server and makes it
difficult to run the agent in unsupported/unknown webservers.  Puma is a
current good example of this.

If the agent is in an environment (e.g. production) where monitor mode
is enabled it will now attempt to start.  I'm planning to add logic in a
later commit to prevent the agent from auto-starting in certain cases,
like in an IRB session/console.