RUBY-554 Goodbye, shoulda.
No tests were harmed during creation of this commit. These tests are
some of the oldest we have, and some are of dubious utility at this
point, but the goal here was mainly to remove our usage of shoulda and
make our test names a bit more sane, with a bit of opportunistic cruft
cleanup on the way.