RUBY-1059 Add get_transaction_name API
I've added a get_transaction_name API which is symmetrical to
set_transaction name.

A very common use case for set_transaction_name is to take the default
name and alter it in some minor way, for example adding the mime type
(i.e. .json, .xml, .html).  This provides an easy way to read the
current name out so it can be tweaked to perfection.