RUBY-1030 Don't call Sinatra::Base.process_route out of context
Previously the agent would call process_route to try and determine what route would run.
This is perilous as conditions might not expect it, and routes might pass which would
change the final route we actually run.

Now instead the agent will capture the route names as they are processed, but only apply
it to the transaction at route_eval, when we actually run. If we pass to another, we'll
capture again and call route_eval again, resetting the transaction name with it.