RUBY-1057 JRuby fixes for Sequel instrumentation
Sqlite on JRuby doesn't report back adapter, so we don't record the RemoteService metric.
Issues with the explain durations coming back as 0.0 not triggering the plan generation.
Additionally were receiving 'ThreadError: Mutex relocking by same thread' on adding the
extension to the database because of our own synchronize block around it.