RUBY-1090 Make AgentHooks middleware idempotent
Even if we get multiple additions of it in the chain, only fire the before/after events once.
This is a concern as Sinatra clients could have added things manually at the Rack layer, and
our auto-injection will now put the middleware in again. Since we can't strictly control
whether the middleware is somewhere in the stack, we can only do our best to behave when we're
duplicated