RUBY-1109 - Move constant parsing to config manager, combine config, refactor
The constant parsing is now handled in the configuration manager to prevent the
string from being parsed for each noticed error.

Configuration for the strip_exception_messages feature is now a hash with keys
'enabled' and 'whitelist', where the whitelist is a comma separated string of
exceptions that allow messages.

When the whitelist contains an undefined constant a warning is written to the
log with the constant name.