RUBY-1113 Fix broken CPU, memory samplers for forking servers.
We recently started resetting the StatsEngine instance from
NewRelic::Agent::Agent#after_fork.

Because the list of harvest-time samplers was attached to the
StatsEngine, this meant that no harvest time samplers (notably: CPU
and memory samplers) would work from forking servers such as Unicorn
and Passenger.

This commit moves the list of samplers into a SamplerManager that is
attached directly to the Agent, and hooks into the harvest cycle
via a new :before_harvest event.