RUBY-1119 Add excon instrumentation support
Excon has built-in support for an ActiveSupport::Notifications-like interface,
but I don't think it's quite the right thing to use for our instrumentation.
There are two ways to configure an instrumentor that will recieve
request/response lifecycle events: by changing the default instrumentor in
Excon.defaults or by specifying the :instrumentor key when creating a new
connection.

The former seems inappropriate for us in that setting the default instrumentor
exposes us to the possibility of a customer's app (or another library that their
app uses) changing the default instrumentor after us, causing us to lose
instrumentation (or us doing same to someone else).

The latter (setting :instrumentor explicitly when creating reqeusts) also seems
a bad fit for our needs, since it's opt-in, and must be done for each request
that's created.

So, instead of using the AS::Note instrumentation stuff, I opted to use Excon's
middleware infrastructure to inject a New Relic middleware to wrap requests and
responses. This has the advantage of allowing us to inter-operate more easily
with other libraries / apps that want to do Excon instrumentation of their own.