Check for ObjectSpace and helper methods to define/undefine constants safely
object_space_enabled? will return true if we have access to ObjectSpace and
false otherwise, with special attention paid to JRuby as it defines ObjectSpace
so it can raise errors about ObjectSpace not being available.

define_constant will define a constant with a supplied implementation for the
duration of the test, then reimplement the original implementation.

undefine_constant will undefine a constant for the duration of a test and
redefine it afterwards.