Capture stderr in the OutputCollector in multiverse to ease debugging multiverse tests
We use IO.popen to spawn each child environment in a multiverse suite,
and this works fine for capturing stdout of the processes that are
executing tests, but it doesn't work for capturing stderr.

I suspect that some of the mysterious failures we see in multiverse
where there's no apparent error message would be less mysterious if we
could actually capture the stderr of the test processes.

This change accomplishes that by redirecting stderr -> stdout as the
first step when running a child environemnt of a multiverse suite.