No worker thread in multiverse
Although we haven't established it for sure, we suspect that the worker thread might
be the cause of some of our multiverse instability. Since multiverse does its own collection
synchronously within the tests, we don't need the harvest thread doing too.

This also adds a before/after file-based inclusion mechanism in the child multiverse
processes.