RUBY-1149 Run suite environments in parallel
This changes the suite execution for multiverse so it starts up each of the
generated environments in a suite in parallel. It uses a renamed Gemfile for
each of them so they can safely run at the same time. Output handling is also
a bit different, as we don't interleave the test runs, but wait until each
finishes to show the result.