RUBY-1159 Don't factor agent errors into queue limit
We limit noticed errors to a cap of 20, but we don't want agent errors that we are potentially
capturing to trump app errors. Neither do we want to exclude agent errors because the app's
filled things up already.

Since the agent errors mechanism has its own per-class cap (and we'll only be using it for
very specific diagnostic situations), let the agent errors past the queue limit and don't
count them in the overall total.