RUBY-1130 Refactor DJ instrumentation to fix instrumentation for 2.x
This change guts the previous delayed_job_injection instrumentation
payload, and instead moves to an approach where all DJ instrumentation
lives in the delayed_job instrumentation payload.

Because of a pecularity of DJ's startup sequence, the instrumentation
is installed in two phases: in phase 1, we hook the first phase is
triggered by the normal agent startup sequence, and hooks the
Delayed::Worker#initialize method in order to do some initialization
and install the second phase.

The second phase actually wraps the Delayed::Job#invoke_job method in
a transaction tracer. This needs to be deferred until a second phase
because the Delayed::Job class is not yet available at the time when
phase 1 of the instrumentation is installed.

Aditionally, I've removed our previous wrapper around
Delayed::Job#after_fork, since it should no longer be needed. By the
time Delayed::Worker#initialize is hit, after_fork has already been
invoked. References:

3.x: https://github.com/collectiveidea/delayed_job/blob/master/lib/delayed/command.rb#L106
2.1.x: https://github.com/collectiveidea/delayed_job/blob/v2.1/lib/delayed/command.rb#L102
2.0.x: https://github.com/collectiveidea/delayed_job/blob/v2.0/lib/delayed/command.rb#L102