RUBY-1130 Loosen requirements to identify delayed_job as dispatcher
Some users defer the require of delayed_job until a Rails initializer,
which means that the Delayed::Worker constant may not be defined yet
at the time when we try to detect the dispatcher.

Since we rely on detection of the dispatcher in order to defer startup
of the agent until after DJ daemonizes, this presents a problem. So,
this change loosens the dispatcher detection requirement for
delayed_job to be based solely on the script name ($0).

If this needs to be overridden for some reason, the
NEWRELIC_DISPATCHER environment variable can do the trick.