RUBY-1173 fix excon instrumentation usage with idempotent + expects
When a request made via excon uses the :idempotent and :expects options,
it will be retried automatically if the actual HTTP response code doesn't
match what the caller expected.

This previously created a problem for the Ruby agent's instrumentation of
excon, because we were assuming that request entries (via the request_call
method on our middleware) would always be paired up with exits (via the
response_call or error_call middleware methods).

In actuality, we recieve multiple request_call calls but only a single
response_call / error_call call when a request is retried automatically by
excon.

This change attaches our trace data to the Excon::Connection instance
directly rather than stashing it on the params hash (since the Connection
is reused on a retry but the params hash is not) and checks whether a
given connection is already in the middle of a call before tracing it again.