RUBY-1114 Improve the way we keep track of which exceptions we've seen
The previous approach to keeping track of exceptions that we had already
recorded for the purposes of error counting in the agent would attempt to
write an ivar onto the exception object in order to mark it as seen.

This doesn't work with frozen objects, and raises a warning with certain
kinds of exceptions in JRuby.

The new approach attaches the object_ids of the seen exceptions onto
the current transaction if there is one, and then consults this list
before counting a new error.

Note that the de-duping here is narrower in scope than the previous approach:
it's scoped to the lifetime of the transaction (under the assumption
that the same error object won't be seen across two different transactions).