RUBY-1176 Fix test problem with unstub
To get test to properly replace a method that we stub by default in setup, we were trying to
unstub. This doesn't seem to work right on the version of mocha encountered on the rails23
env locally. Now we just replace out the objects we're testing with a clean slate instead.