RUBY-1132 Use thread-safe ActiveSupport::Notifications API for ActiveRecord instrumentation on Rails 4.0
We use ActiveSupport::Notifications for our Rails 4 instrumentation in three places:
ActiveRecord, ActionController, and ActionView.

The latter two of these use the start/finish callbacks form of
ActiveSupport::Notifications#subscribe, but the first used the simpler 'call'
form (where only one callback with both start and end times attached is
emitted).

The 'call' version of ActiveSupport::Notifications (which is the default) is
not thread-safe. See https://github.com/rails/rails/issues/12069 for details.

This change switches our ActiveRecordSubscriber over to use the start/finish
form of subscription, via the EventedSubscriber class that's also used by
ActionController and ActionView instrumentation. It also ensures that all of
our subscribers are rescue from errors that happen during our callbacks, so
that these errors don't propagate and destroy the request.