RUBY-1183: Remove some prefixes, add some configuration settings, refactor.
Prefixes that do not include an underscore have will not be treated as
configuration settings. Valid prefixes are 'NEWRELIC_' and 'NEW_RELIC_'.

Three previously undocumented configuration settings now have defaults:
  * :license_key
  * :log
  * :omit_fake_collector

Settings that do not have defaults generate info level log messages and are
included in the config as they are read from the environment, as strings.
As a result arbitrary values with the correct prefix will be available in the
environment report.