RUBY-1193 Fix for partial lookup problem
We were previously patching into ActiveView::PartialRenderer#render and doing part of the
work the view does to find our name for the related scope push. Unfortunately, if a partial
is rendered with the format embedded in the string (deprecated but still allowed), then
the call to setup we were making actually mutates data and results in a different partial
path the second time (which probably fails to render).

This commit changes the tactic to instead override the instrument method that's being called
once the identifier for the partial has already been found. The data we need is in the
payload to just look up.