RUBY-1176 Beefing up TransactionSampleBuffer class
Previously the various sample buffers had a lot of similar functionality,
though mildly different in it's own special way. This commit consolidates that
to using the base class TransactionTracer as the primary source for:

* A list of samples
* Managing how many samples to hold onto
* Truncating when necessary (with overrides for the condition and ordering in
  those cases)

This leaves the sample buffer classes a lot slimmer, which is good, and sets
the pattern for others later on to follow. It does mean you need to understand
the base class a bit, but that seems a reasonable tradeoff.