RUBY-1176 Limit TT's through buffers
Now that we have nice buffer classes for holding samples, let them manage the
limits on how many traces each type will be allowed to send. This mildly changes
the limits around re-transmission, but it does continue to limit them to a sane
number and is better tested and clearer.