RUBY-1196 Default transaction name in Rails 4
By default Transaction doesn't set a name set until the close of a transaction.
However for Rails 4 this causes problems if you manually include RUM
header/footers which cause a name freeze before we're ready.

While we don't want to globally set transaction names at the start of
transactions, it is safe for the ActionControllerSubscriber to do so, which
protects us from the manual RUM issues.