RUBY-1196 Default transaction name to non-nil
When we start a transaction in the past we left the name nil so that we could
detect it and set it to the first metric name when stopping the transaction.

That's all well and good until things like RUM cause us to prematurely freeze
the name as nil, and later browser monitoring bombs out not getting the string
it expects.

Instead, we now set the transaction name to (unknown) when it is started, and
the internal checks for name == nil use name_set? instead to see if we should
be overwriting the name.