Supported Versions for the Ruby Agent
This is an initial draft of documenting (and automating processes) around the
gem versions that we support from the Ruby agent.

This includes:

* NewRelic::SUPPORTED_VERSIONS data structure with metadata on each framework
  and Ruby version we support
* Rake task (newrelic:supported_versions) for emitting documentation in either
  plain text or HTML for the current agent

All data at this point is the best I've been able to gather without
exhaustively running really old versions of things, and in general we should be
encouraging people to run modern versions of their gems where possible.