RUBY-1202 Fix for get_agent_commands being called from Resque jobs
The AgentCommandRouter (introduced in 3.6.7) was holding onto its own
reference to the NewRelicService, meaning that when the service reference
from the main agent got swapped out for a PipeService, it didn't see the
change. This resulted in every Resque job hitting the get_agent_commands
collector endpoint upon exit. This change addresses the issue by having
the AgentCommandRouter use the Agent's handle to the service.