Add the ability to pass a timer to performanc tests.
This allows you to do some test-specific setup that you don't want timed
within each test itself, rather than forcing all setup into the setup method.

This allows greater flexibility when writing groups of performance tests that
may share some, but not all setup tasks in common.

This feature is opt-in - to use it, just make your test method take a single
argument, which will be a Timer object. You can then call #measure on the
given Timer and pass in the block of code whose execution you'd like timed
for the test (as opposed to the default of the entire test method).