RUBY-1110 Don't change locking semantics
Prior positioning for this lock changed what we we were actually locking around
subtly. While not certain to cause an issue, better safe than sorry.

Agent#transmit_data is still called as the entry point, but we synchronize
within it and call a (private) method with what used to be transmit_data's
guts.