RUBY-1176 Gather accurate thread count
For whatever historical reason, we were sending the sample count along in the
spot that the collector expects a unique thread count. Since it isn't surfaced
in the UI, this never came up as an issue, but it's totally wrong.

This fixes that by tracking the unique threads we aggregate on in the thread
profile, and using that number in the collector payload.