RUBY-1176 Stop sessions after duration
While we normally rely on the collector to tell us when our X-Ray sesisons are
done, if we lose touch with the collector we also want to stop a session when
we are sure it must be done (based on the local session's start time + the
duration that the collector told us).

This prevents us from potentially unbounded growth when we can't get the
message an X-Ray should be finished, and can't ship the data off to the
collector.