Fixes for Rubinius 2.0 GC support
Previous versions of the agent had specific support for GC metrics in Rubinius.
These were largely untested, and appear to now be unnecessary with rbx-2.0.0 as
it implements the Core ::GC::Profiler class.

This commit also revises the names of the small adapter classes used for the
various flavors of GC (REE, Core and legacy Rubinius)