Annotate methods that are part of our public API with @api public for YARD
This is a step towards having a more well-defined API for the agent.
Additionally, I've added a .yardopts file that (I think) will be used when
generating documentation for rubydoc.info. The default set of options in this
file will make it so that only methods / classes / modules that are explicitly
marked as public will appear in the public-facing documentation.