RUBY-1212 Place maximum bound on TT's to buffer
Each of the sample buffers we use for TT's will now hold a maximum of 20 items
a piece. This keeps users from tweaking settings and consuming too much memory
for little benefit.

All gathered samples in the buffers will continue to be sent to the collector,
but this should limit the damage a client can do to themselves with the
settings.