RUBY-1219 Don't defer Resque startup if can't fork
We typically defer startup of the agent until the before_first_fork hook in
Resque so we can catch if the master process daemonizes. On JRuby we don't
install the before_first_fork hook, though (since we can't fork), but we also
defer startup.

Instead, skip deferring if we're on a platform that can't fork.