RUBY-1226 Add coercion and validation to custom params in events
Since custom parameters are user provided values that can be of any
arbitrary type we coerce them into a format that's acceptable for this
use case.

We accept a Ruby Hash and will return a new Hash where all of the keys
have been converted to strings.  As values we only allow Strings,
Floats, Integers, and Symbols (which are converted to strings when
they're later converted to JSON).  Any values of other type (e.g. Hash,
Array, any other class) are discarded. Their keys are also removed from
the results hash.

I'm not super happy with the #event_params method name but haven't yet
thought of anything better.