RUBY-1226 Don't crash if bad argument is passed to add_custom_parameters
I was getting errors when I accidentally passed a String to
add_custom_parameters instead of a Hash.  Since we're going to be using
this API more heavily in the future it seems like a good idea to log
these argument errors as warnings instead of just crashing the request.