RUBY-1232 Config for custom params in JS script
By default, custom parameters will not be included in the JS agent's script on
the page, but that can be controlled with a config key.

Inclusion of custom parameters in the JS agent block will also be influenced by
the other analytics enabling config settings, both globally and for
transactions.