Optionally serialize multiverse tests to avoid swapping on CI servers
Environments in each multiverse suite will still be executed in parallel by
default, but setting the environment variable SERIALIZE will cause the
environemnts to be run sequentially instead. The plan is to set this in CI and
measure the performance difference. A quick test on ruby-build-box-1 with the
sequel multiverse suite indicated that test serialization would be helpful
(the suite finished in 7.5 minutes with SERIALIZE=1, didn't finish after ~20min
without serialization).