RUBY-1243 Require encoding_key for cross app
In addition to making sure that we have the cross_process_id, we also are
gating the cross app features on having a non-zero length encoding key.

The obfuscation is permissive to nil's and empty strings, so this prevents us
from sending non-obfuscated values across the wire if the collector were ever
to send us a cross_process_id but no encoding_key.