RUBY-1166 Restart worker loop after daemonize
This first step hooks the start_transaction event, and will check whether we
have tried to start up the worker thread within the current process yet. If we
haven't, then we'll call to NewRelic::Agent.after_fork to get that doing.

NewRelic::Agent.after_fork also gates on the same flagging to ensure that we
don't get into a race to start more than one worker thread.

In general this should solve our problems with any daemonizing processes. The
only slight hang-up is that it only triggers after a transaction in the child
processes, but that is vastly preferrable to getting no data or requiring
client or per-library intervention.