RUBY-1244 Don't double trace in Net::HTTP
If a connection isn't started yet, Net::HTTP will end up calling recursively
into request twice in the course of making the call.

We previously handled this by only tracing request after the connection was
started, but in certain timeout scenarios, this fails to get an external trace
at all. It also missed out on the connection startup time, not attributing it
to the HTTP call.

Now we trace the outer-most request, but stop any further tracing within
Net::HTTP so the potential inner call is ignored.