RUBY-1244 Adding unit tests for CrossAppTracing
All the testing for CAT was in multiverse. While those tests are great and need
to continue being extended (especially in the face of how we use CAT from
multiple HTTP libraries), there are times when it'd be nice to more directly
unit test individual methods, this makes a home for those.