RUBY-1237 Add normalize_string to JsonMarshaller to handle string encodings
On platforms that support it, we use the JSON implementation from Ruby's stdlib
to serialize data to be sent to the collector.

The JSON gem README sayeth:

  "All strings, that are to be encoded as JSON strings, should be UTF-8 byte
  sequences on the Ruby side."

Unfortunately, strings come to the agent through a variety of paths, and are
sometimes either not encoded as UTF-8, or are tagged as being UTF-8 when they
do not actually contain valid UTF-8 byte sequences.

In this case, we have several options:

  1. Just drop the strings
  2. Munge the strings so that they can be represented as UTF-8
  3. Use a non-JSON serialization format that can represent all Ruby Strings

Up until now, we've been doing option 3 (using the Marshal format when encoding
to JSON fails). We want to stop doing that. Among the remaning two options, #2
is friendlier.

Hence, NewRelic::Agent::NewRelicService::JsonMarshaller#normalize_string

This method takes a String, and returns either that same string (if it was
already suitable for encoding to JSON), or a munged copy of the input string
(without modifying the original).

The munging that we do upon encountering a String that cannot be cleanly
converted to UTF-8 consists essentially of treating the string as if it were
ISO-8859-1 (in which any byte sequence is valid), and then encoding the result
to UTF-8.

It's notable that this munging is reversible, assuming the correct original
encoding is known.