RUBY-1237 Handle bad strings in double-encoded object graphs
Sometimes we do this thing where we take an object graph that's ready to
be serialized, pick out certain non-atom fields, and decide to encode those
fields as base64-encoded, gzip-compressed JSON strings before actually
serializing the outer object graph.

This happens via the to_collector_array methods implemented on a few different
classes in the agent.

Prior to this change-set the encoding normalization logic in the JsonMarshaller
did not correctly handle the case where one of the strings inside of one of
these double-encoded blogs had an invalid encoding. It also did not correctly
handle the case of custom classes implementing to_collector_array.

Previously, we would walk the object graph before passing it through
JsonMarshaller#prepare. This was incorrect, because many of the objects in the
graph had not yet been transformed to the representations we were going to use
for serialization - they were instead instances of classes that implemented
to_collector_array.

Now, we pass the object graph to be serialized through JsonMarshaller#prepare
first, and then through normalize.