RUBY-1237 Funnel all JSONing through NewRelic::JSONWrapper
Previously, we had a mixture of direct references to JSON.dump/load and
NewRelic.json_dump/json_load scattered throughout the code.

Now, all JSON encoding/decoding passes through NewRelic::JSONWrapper.dump/load.

These methods either wrap the standard-library (or gem) JSON implementation, or
fall back to OkJson on older Ruby versions. An optional :normalize option
triggers the encoding normalization behavior that we want for serializing data
to be sent to the collector, and is only supported on Ruby 1.9+.

On newer Ruby versions, grabbing the JSON.dump/load methods at agent startup and
caching them for us from the JsonMarshaller in the NewRelicService should make
it so we're more likely to get the JSON implementation we're expecting, and not
be subject to potential bugs in alternate implementations that hijack the JSON
constant.