RUBY-1233 Whitelist collection name from selector
Previously we relied on getting the first item from the selector hash to get
the collection for $cmd's in Mongo. This didn't work on 1.8.7 because the
hashes aren't ordered.

Instead we whitelist the known commands that we look for in the selector and
anything we don't know will end up recording a metric including $cmd in the
name instead.